refactor: modularise files (#1772)
This is basically a mechanical change to separate `files-regular.js` into multiple modules - easier to read, debug and maintain.

License: MIT
