+++ src/core/components/files-regular/cat-readable-stream.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const toStream = require('pull-stream-to-stream')
+
+module.exports = function (self) {
+  return (ipfsPath, options) => toStream.source(self.catPullStream(ipfsPath, options))
+}
