+++ src/core/components/files-regular/ls-readable-stream.js
@@ -1,0 +1,9 @@
+'use strict'
+
+const toStream = require('pull-stream-to-stream')
+
+module.exports = function (self) {
+  return (ipfsPath, options) => {
+    return toStream.source(self.lsPullStream(ipfsPath, options))
+  }
+}
