+++ test/core/node.js
@@ -1,6 +1,7 @@
 'use strict'
 
 require('./circuit-relay')
+require('./files-regular-utils')
 require('./name')
 require('./key-exchange')
 require('./pin')
